Added the basic shell structure
Parsing the line and splitting it into arguments

I like more the approach in embedded in which you process the next argument
instead of splitting into an array.

Two main functions created and their unit tests

Still all the code to handle the back and forwards is missing.

Modified the flush mechanism to be able to flush if the process
has completed.

Signed-off-by: Sergio Martinez <sergio.martinez.rodriguez@intel.com>